Remove "Depends:" field from DESCRIPTION
The dependency on R >= 3.5 seemed excessive to me. [A few months ago I did some asking around](https://community.rstudio.com/t/determining-which-version-of-r-to-depend-on/4396) on what the recommended approach on R version dependency is, and [Hadley summarized it as follows:](https://community.rstudio.com/t/determining-which-version-of-r-to-depend-on/4396/10)

> Now, I think you should only explicitly require an R version under two circumstances:
> - You are required to by R CMD check
> - You have checked that your package works on multiple versions of R and you want to assert that. (This is now easy on travis)

I have executed R CMD check locally after removing the DESCRIPTION field, and I assume it would be fairly easy to add `oldrel` or earlier versions to your `.travis.yml` and see what happens. Putting the dependency back in is always possible anyway.